Worked On 1. Claims View All, Customer name, Removed Email ID from List by adding another function to provide only display name of a user 2. Aligned the header for all table to left and Text is made bold 3. Removed SNO column for table (Role, Operation, etc..) 4. Aligned the header to left by default and center for Action, Date, Percentage 5. Incereased the width of description column viewed in the viewAll table 6. Added service Zip code in Service provider > contractor > view All table 7. Added -NA- For Service in Zip code and Type if value is not available in service provider > contractor > view all 8. Expand all / collapse all for Service provider contractor , Adjuster and claim 9. Moved few common text together, 10. Added some cosmetic changes for claim view one 11. Provided space between header text and menu in all pages 12. Provided delete support by adding deleted flag and corrected the flag data on following 	Contractor, Permissions, Tasks, Budget, Project, Docs, Issues, Partner, Notes 13. Header added for Contractor > Trades and Sub trades 14. Discount trades and sub trades heading text are populated from Master DB. It was throwing error which is corrected. 15. Moved accordion init, expand and collapse accordion function to utils, common place and cleared the code in all modules 16. expand all and colaps all for subrogation view, Util function for expand all and collapse all 17. Added Permission as admin for all the security module view forms. 18. Permission Check on db updates and deployment the site 	- User 	- Role 	- Function 	- Operation 	- Data Filter 	- Permission

+++ application/controllers/security/permissions.php
@@ -15,6 +15,18 @@ class Permissions extends CI_Controller {
 	}
 
 	public function getComponentInfo() {
+		$response = array(
+			'status'	=> "error"
+		);
+		/* Get Role ID and Role Display String*/
+		list($role_id, $role_disp_name) = $this->permissions_lib->getRoleAndDisplayStr();
+
+		if($role_disp_name != "admin") {
+			$response["message"] 			= "No permission to execute this operation";
+			print_r(json_encode($response));
+			return false;
+		}
+		
 		$this->load->model('security/model_permissions');
 
 		$getParams = array(
@@ -30,6 +42,16 @@ class Permissions extends CI_Controller {
 	}
 
 	public function inputForm() {
+		/* Get Role ID and Role Display String*/
+		list($role_id, $role_disp_name) = $this->permissions_lib->getRoleAndDisplayStr();
+
+		if($role_disp_name != "admin") {
+			$no_permission_options = array(
+				'page_disp_string' => "create permission"
+			);
+			echo $this->load->view("pages/no_permission", $no_permission_options, true);
+			return false;
+		}
 		$this->load->model('security/model_permissions');
 
 		$permissionId = $this->input->post('permissionId');
@@ -58,6 +80,18 @@ class Permissions extends CI_Controller {
 	}
 
 	public function getPageForType() {
+		$response = array(
+			'status'	=> "error"
+		);
+		/* Get Role ID and Role Display String*/
+		list($role_id, $role_disp_name) = $this->permissions_lib->getRoleAndDisplayStr();
+
+		if($role_disp_name != "admin") {
+			$response["message"] 			= "No permission to execute this operation";
+			print_r(json_encode($response));
+			return false;
+		}
+
 		$this->load->model('security/model_permissions');
 
 		$type = $this->input->post('type');
@@ -108,6 +142,18 @@ class Permissions extends CI_Controller {
 	}*/
 
 	public function getPermissions() {
+		$response = array(
+			'status'	=> "error"
+		);
+		/* Get Role ID and Role Display String*/
+		list($role_id, $role_disp_name) = $this->permissions_lib->getRoleAndDisplayStr();
+
+		if($role_disp_name != "admin") {
+			$response["message"] 			= "No permission to execute this operation";
+			print_r(json_encode($response));
+			return false;
+		}
+
 		$this->load->model('security/model_permissions');
 
 		$user_id = $this->input->post('user_id');
@@ -143,6 +189,18 @@ class Permissions extends CI_Controller {
 	}
 
 	public function setPermissions() {
+		$response = array(
+			'status'	=> "error"
+		);
+		/* Get Role ID and Role Display String*/
+		list($role_id, $role_disp_name) = $this->permissions_lib->getRoleAndDisplayStr();
+
+		if($role_disp_name != "admin") {
+			$response["message"] 			= "No permission to execute this operation";
+			print_r(json_encode($response));
+			return false;
+		}
+
 		$this->load->model('security/model_permissions');
 		$response = $this->model_permissions->setUserPermission();
 
@@ -150,6 +208,18 @@ class Permissions extends CI_Controller {
 	}
 
 	public function deleteRecord() {
+		$response = array(
+			'status'	=> "error"
+		);
+		/* Get Role ID and Role Display String*/
+		list($role_id, $role_disp_name) = $this->permissions_lib->getRoleAndDisplayStr();
+
+		if($role_disp_name != "admin") {
+			$response["message"] 			= "No permission to execute this operation";
+			print_r(json_encode($response));
+			return false;
+		}
+
 		$this->load->model('security/model_permissions');
 		$response = $this->model_permissions->deleteRecord();
